+++ docs/changelog.md
@@ -5,6 +5,8 @@
 * **New**
     * Added `dusty doctor` command to attempt to fix some common VirtualBox networking problems
     * Added a custom 502 page for apps which use host forwarding. This lets you see the status and logs of your container while you're waiting for its HTTP service to come online
+* **Misc**
+    * Implemented a [networking fix](https://github.com/docker/machine/pull/3112) the Dusty team implemented in Docker Machine within Dusty
 
 ## 0.7.0 (February 18, 2016)